+++ src/paymentService/CashPaymentStrategy.java
@@ -2,7 +2,7 @@ package paymentService;
 
 public class CashPaymentStrategy implements PaymentStrategy{
     @Override
-    public void pay(int amount) {
+    public void pay(long amount) {
         System.out.println(amount + " Paid via Cash");
     }
 }
