+++ src/paymentService/CreditCardPaymentStrategy.java
@@ -3,7 +3,7 @@ package paymentService;
 public class CreditCardPaymentStrategy implements PaymentStrategy {
 
     @Override
-    public void pay(int amount) {
+    public void pay(long amount) {
         System.out.println(amount + " paid via credit card");
     }
 }
