+++ src/paymentService/NetBankingPaymentStrategy.java
@@ -2,7 +2,7 @@ package paymentService;
 
 public class NetBankingPaymentStrategy implements PaymentStrategy {
     @Override
-    public void pay(int amount) {
+    public void pay(long amount) {
         System.out.println(amount + " paid via netbanking");
     }
 }
